+++ src/gui/main_window.cpp
@@ -151,6 +151,20 @@ void MainWindow::toggleStatusBar (bool visible)
 }
 
 
+//-------------------------------------------------------------------
+void MainWindow::setOverviewActionEnabled (bool enabled)
+{
+    overview_action_->setEnabled (enabled);
+}
+
+//-----------------------------------------------------------------------------
+void MainWindow::toggleOverview (bool visible)
+{
+    QSettings settings;
+    settings.setValue ("MainWindow/overview", visible);
+    emit overviewVisibilityChanged (visible);
+}
+
 //-----------------------------------------------------------------------------
 void MainWindow::addBackgroundProcessToStatusBar (QString name, int max)
 {
@@ -239,9 +253,16 @@ void MainWindow::initMenus (QSharedPointer<ApplicationContext> application_conte
     toggle_status_bar->setChecked (settings.value ("MainWindow/statusbar", true).toBool());
     connect (toggle_status_bar, SIGNAL(toggled(bool)), this, SLOT(toggleStatusBar(bool)));
 
+    overview_action_ = new QAction (tr("Overview"), this);
+    overview_action_->setCheckable (true);
+    overview_action_->setChecked (settings.value ("MainWindow/overview", true).toBool());
+    overview_action_->setEnabled (false);
+    connect (overview_action_, SIGNAL(toggled(bool)), this, SLOT(toggleOverview(bool)));
+
     view_menu_ = menuBar()->addMenu(tr("&View"));
     view_menu_->addAction(file_toolbar_->toggleViewAction());
     view_menu_->addAction(toggle_status_bar);
+    view_menu_->addAction(overview_action_);
 #ifndef Q_OS_MACOS
     toggle_menubar_ = new QAction(tr("Menubar"), this);
     toggle_menubar_->setCheckable(true);